Updated PracticalSocket and moved it to middleware/libs
